Add assertions to code to if assumptions hold
Several methods have implicit assumptions of specific properties in
parameters and variables.

It will be better to check if these assumptions hold, before moving on
to run the other parts of the code. This helps the debugging process
since the code will stop running at the point where one assumption made
does not hold true.