Handle all exceptions thrown
Some exceptions thrown are not handled, which causes the program to
become unusable.

When exceptions are caught in the UI and error messages are returned as
text messages from Bob, users will have a better understanding of the
error that has occurred.

Let's update the methods in Bob.java and Ui.java to handle all
exceptions by returning the error messages as Strings.